+++ packages/salesforcedx-vscode-core/src/messages/i18n.ts
@@ -532,12 +532,11 @@ export const messages = {
   EmbeddedServiceMenuSettings: 'Embedded Service Menu Settings',
   CallCoachingMediaProvider: 'Call Coaching Media Providers',
 
+  conflict_detect_execution_name: 'Conflict Detection',
   conflict_detect_error:
     'An error was encountered during conflict detection. %s',
-  conflict_detect_retrieve_org_source:
-    'Conflict Detection: retrieving org source',
-  conflict_detect_convert_org_source:
-    'Conflict Detection: converting org source',
+  conflict_detect_empty_results:
+    'Conflict Detection failed to retrieve org file state',
   conflict_detect_conflicts_during_deploy:
     'Conflicts are detected while deploying metadata. Select Override Conflicts to proceed or Cancel to view the conflicts.',
   conflict_detect_conflicts_during_retrieve:
@@ -561,9 +560,12 @@ export const messages = {
   conflict_detect_diff_command_title: 'Compare Files',
 
   force_source_diff_text: 'SFDX: Diff File Against Org',
+  force_source_diff_remote_not_found:
+    'Selected file or directory was not found in the org',
   force_source_diff_unsupported_type:
     'Diff for this metadata type is currently not supported',
   force_source_diff_title: '%s//%s ↔ local//%s',
+  force_source_diff_folder_title: '%s - File Diffs',
   force_source_diff_command_not_found:
     'To run this command, first install the @salesforce/sfdx-diff plugin. For more info, see [https://forcedotcom.github.io/salesforcedx-vscode/articles/user-guide/source-diff](https://forcedotcom.github.io/salesforcedx-vscode/articles/user-guide/source-diff).',
   beta_tapi_mdcontainer_error: 'Unexpected error creating metadata container',
@@ -589,5 +591,6 @@ export const messages = {
   force_sobjects_refresh: 'SFDX: Refresh SObject Definitions',
   sobject_refresh_all: 'All SObjects',
   sobject_refresh_custom: 'Custom SObjects',
-  sobject_refresh_standard: 'Standard SObjects'
+  sobject_refresh_standard: 'Standard SObjects',
+  force_source_diff_components_not_in_org: 'Selected components are not available in the org'
 };