+++ packages/salesforcedx-vscode-core/test/vscode-integration/conflict/metadataCacheService.test.ts
@@ -23,10 +23,12 @@ import * as shell from 'shelljs';
 import {
   MetadataCacheExecutor,
   MetadataCacheResult,
-  MetadataCacheService
+  MetadataCacheService,
+  PathType
 } from '../../../src/conflict/metadataCacheService';
 import { stubRootWorkspace } from '../util/rootWorkspace.test-util';
 import sinon = require('sinon');
+import { SfdxPackageDirectories } from '../../../src/sfdxProject';
 
 describe('Metadata Cache', () => {
   describe('Metadata Cache Executor', () => {
@@ -124,15 +126,20 @@ describe('Metadata Cache', () => {
     const usernameOrAlias = 'admin@ut-sandbox.org';
     const PROJECT_DIR = path.join(PROJ_ROOT, 'meta-proj2');
     let workspaceStub: sinon.SinonStub;
+    let packageStub: sinon.SinonStub;
     let service: MetadataCacheService;
 
     beforeEach(() => {
       service = new MetadataCacheService(usernameOrAlias);
+      packageStub = sinon
+        .stub(SfdxPackageDirectories, 'getPackageDirectoryFullPaths')
+        .resolves([]);
       workspaceStub = stubRootWorkspace(PROJECT_DIR);
     });
 
     afterEach(() => {
       service.clearCache();
+      packageStub.restore();
       workspaceStub!.restore();
       shell.rm('-rf', PROJECT_DIR);
     });
@@ -192,6 +199,26 @@ describe('Metadata Cache', () => {
       expect(components.size).to.equal(14);
     });
 
+    it('Should find components using a manifest', async () => {
+      const projectPath = path.join(PROJECT_DIR, 'src');
+      const manifestPath = path.join(
+        TEST_ASSETS_FOLDER,
+        'proj-testdata',
+        'manifest',
+        'one-class.xml'
+      );
+
+      // populate project metadata
+      const projectZip = new AdmZip();
+      projectZip.addLocalFolder(TEST_DATA_FOLDER);
+      projectZip.extractAllTo(projectPath);
+
+      service.initialize(manifestPath, PROJECT_DIR, true);
+      const components = await service.getSourceComponents();
+
+      expect(components.size).to.equal(1);
+    });
+
     it('Should return cache results', async () => {
       const projectPath = path.join(PROJECT_DIR, 'src');
       const cachePath = service.getCachePath();
@@ -211,7 +238,7 @@ describe('Metadata Cache', () => {
 
       expect(cache).to.not.equal(undefined);
       expect(cache?.selectedPath).to.equal(projectPath);
-      expect(cache?.selectedIsDirectory).to.equal(true);
+      expect(cache?.selectedType).to.equal(PathType.Folder);
       expect(cache?.cachePropPath).to.equal(
         path.join(cachePath, 'prop', 'file-props.json')
       );
@@ -240,6 +267,7 @@ describe('Metadata Cache', () => {
   });
 
   async function handleCacheResults(
+    username: string,
     cache?: MetadataCacheResult
   ): Promise<void> {}